pyproject.toml
[project]
name = "celerentis"
version = "0.1.0"
description = "AI-assisted, template-driven deck generator for M&A/PE/VC."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
dependencies = [
  "python-pptx>=0.6.23",
  "pydantic>=2.7",
  "pyyaml>=6.0.2",
  "typer>=0.12",
  "rich>=13.7",
  "pandas>=2.2",
  "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.5",
  "black>=24.8",
  "pre-commit>=3.8",
  "types-PyYAML",
  "types-requests",
]

[project.scripts]
celerentis = "celerentis.cli:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["celerentis*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
src = ["src", "tests", "examples"]

[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["celerentis"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
mypy_path = "src"
files = ["src", "tests"]


[tool.pytest.ini_options]
addopts = "-q --cov=celerentis --cov-report=term-missing"
testpaths = ["tests"]
